Rename package according to our decision
