+++ server/src/main/java/com/lunchtime/controllers/FeedbackController.java
@@ -2,7 +2,7 @@ package com.lunchtime.controllers;
 
 import java.util.List;
 import com.lunchtime.models.Feedback;
-import com.lunchtime.service.implementation.FeedbackServiceImplement;
+import com.lunchtime.service.impl.FeedbackServiceImplement;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
